Update patch set 3
Patch Set 3:

> I have seen projects replacing them self with themselves due to having a circular dependency, e.g:
> 
>     replace github.com/gopher/mod => .
> 
> Does this block this so such projects can no longer work, or is there some other fix so this is no longer necessary.
> 
> (Yes, such circular dependencies are probably a bad idea, but some people just insist on doing them regardless)

Segev, this CL will report an error when the main module replaces itself or replaces another module with itself.

I don't think this kind replacement should be necessary, but please comment on golang.org/issue/34417 if you're running into a problem that's difficult to solve without this, and we'll try and recommend another solution.

Patch-set: 3