Update patch set 1
Patch Set 1:

One possibility for getting back the rest of the CPU time regression is throwing a global cache at the front of tconv, and getting rid of the InternString call. I have it implemented, but need to benchmark well. Hopefully this CL gets us close enough to where we were for 1.14, and we can revisit for 1.15.

Patch-set: 1
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>