Update patch set 5
Patch Set 5:

(4 comments)

Thanks for the review!  Yes, as you said a big part of the perf improvement is from doing the physical-to-logical conversion only after stack deduplication.  But also, using the frame pointers means we have a very quick loop instead of running the complex code in gentraceback() (which requires calling findfunc() even to get just the physical pcs).  I've added a comment to this effect in traceStackID() and in the CL description.

Patch-set: 5