Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> You're right, with --alloc_space things look quite different.
> The default is --inuse_space. I don't really understand what that even means. Inuse when? Average over all the GCs?

In use at the moment that you ask for the profile. that is helpful when profiling running servers. it is unhelpful when profiling the compiler or a testing benchmark, in which case it is profiling what is in use just before the program exits. 

it’s not just you that’s confused. see https://github.com/golang/go/issues/24443 and many other results if you search for inuse_space in the issue tracker. iirc it got better with the introduction of an “allocs” profile that tells pprof to use alloc_space. I thought I switched the compiler profiling to use that, but apparently I didn’t in the end. there’s some discussion in https://github.com/golang/go/issues/25096.




> 
>       flat  flat%   sum%        cum   cum%
>  5812.34kB 67.69% 67.69%  6436.54kB 74.96%  cmd/compile/internal/gc.tconvNew
> -1146.48kB 13.35% 54.34% -1750.39kB 20.39%  cmd/compile/internal/gc.typefmt
> 
>          .          .    668:func tconvNew(t *types.Type, flag FmtFlag, mode fmtMode) string {
>     2.81MB     4.92MB    669:	b := bytes.NewBuffer(make([]byte, 0, 64))
>          .   624.20kB    670:	tconv2(b, t, flag, mode, nil)
>          .   773.44kB    671:	return b.String()
>          .          .    672:}
>  
> Mystery solved.

Patch-set: 2