Update patch set 8
Patch Set 8:

I find some references in fmt package doc:
"precision sets the number of places after the decimal, if appropriate, except that for %g/%G precision sets the maximum number of significant digits (trailing zeros are removed). For example, given 12.345 the format %6.3f prints 12.345 while %.3g prints 12.3."
And for complex:
"For complex numbers, the width and precision apply to the two components independently and the result is parenthesized, so %f applied to 1.2+3.4i produces (1.200000+3.400000i)."
So the tests are wrong, could u confirm this? Then I can fix them, thanks!

Patch-set: 8