Update patch set 8
Patch Set 8:

The tests are correct, the new code is broken. In other words, the failures are catching a bug in the new code and should not be changed.

I tried these formats in C, which is the reference for this stuff, and it agrees with the existing tests:

`
% cat x.c
#include <stdio.h>
int main(void) {
	printf("%#g\n", 1e-10);
	printf("%#g\n", 0.);
	printf("%#12.5g\n", 0.);
	printf("%#12.5g\n", 100000.);
	printf("%#12.5g\n", 1.23e6);
}
% gcc x.c
% a.out
1.00000e-10
0.00000
      0.0000
  1.0000e+05
  1.2300e+06
% 
`

Patch-set: 8
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>