Update patch set 5
Patch Set 5:

> > Alex, why do you disagree with this change?  It appears to make Windows act more like Unix with regard to whether an error should be treated as temporary or not.
> 
> I don't see how this change makes Windows act more like Unix. Do you have some example to demonstrate that?

I don't have a runnable example.  The theoretical example would be a program that opens a network connection, gets an error, and calls Temporary to see whether it should try again or whether it got a permanent error.  The error WSAENOBUFS and WSAEMFILE seem to me to be temporary errors in this sense.  In particular, the Unix error EMFILE, which corresponds to the Windows error WSAEMFILE, is already considered to be temporary in Errno.Temporary in syscall/syscall_unix.go.  So this change seems to make Errno.Temporary more similar across Windows and Unix systems.

Patch-set: 5
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>