cmd/internal/obj/riscv: use signed immediates for U-instructions
On RISCV64, the U-instructions (AUIPC and LUI) take 20 bits, append 12 bits
of zeros and sign extend to 64-bits. As such, the 20 bit immediate value is
signed not unsigned.

Note that the branch and jump instructions change since we are currently
really testing the AUIPC rather than the JALR (which will be fixed later).

Updates #27532

Change-Id: I725215a1dc500106dbfdc0a4425f3c0b2a6f411e