cmd/compile: during rulegen, pre-substitute ops
We now have encoded that some values will be lowered simply
by changing the ops.

If the pre-lowering op occurs in the match part of a rule,
then successful application of that rule depends on evaluation
order: Did the rule get applied before or after the op change?
That's undesirable; since we can detect it, we'd like to refuse it.
Unfortunately, we can't yet, since PPC64 has such rules,
and removing them makes code generation worse.
I plan to fix the situation and enable the check.

If the pre-lowering op occurs in the result part of a rule,
perhaps for readability reasons, we can skip a pass through
the rules file and use the post-lowering op directly in the rule.

Change-Id: I2908920546d89deddcf558b60ea46685204adb91