Update patch set 2
Patch Set 2:

One argument against this long ago was that it creates a performance cliff between two identical-looking operations on different values.

That argument was from long ago when the compiler had very few magic optimizations, though, and I feel like the compiler today is much less predictable (for a human guessing what the codegen might be) than it used to be, so it's possible such arguments are now moot.

Patch-set: 2