Update patch set 12
Patch Set 12:

My understanding is that for the true zero (0) all digits are significant, regardless of their position with regards to the decimal point; e.g. in "0", "0.0", and "0.00". It's generally true that moving the decimal point (i.e. multiplying or dividing by 10^n) doesn't change the significance of the digits, and it holds for the zero as well. Therefore, "%#.4g" should print "0.000", not "0.0000".

This is also consistent with the behavior of C, which states (taken from ISO/IEC 9899:TC2):

> e, E:
If the value is zero, the exponent is zero.

> g, G:
Let P equal the precision if nonzero, 6 if the precision is omitted, or 1 if the precision is zero.
Then, if a conversion with style E would have an exponent of X:
— if P > X ≥ −4, the conversion is with style f (or F) and precision P − (X + 1).
— otherwise, the conversion is with style e (or E) and precision P − 1.

For the zero value, X is 0, so the conversion is with style f (or F) and precision P − 1, so "%#.4g" is equivalent to "%#.3f".

Patch-set: 12
CC: Gerrit User 34942 <34942@62eb7196-b449-3ce5-99f1-c037f21e1705>