Update patch set 1
Patch Set 1:

> Well, for the 5 past years I'm looking into golang people are struggling to make compiler smarter yet simplest cases like this still do not work. You can check the huge history of the string<->[]byte conversion topic in issue tracking. Why having faster std library by using unsafe is considered so bad? we can remove unsafe usages one by one as compiler improves. Moreover, these places will be an exact pieces telling us where compiler is not so good and worth improving.

Using unsafe in the standard library, outside of specialized areas like runtime, reflect, and syscall, is bad because it shows a deficiency in the language or the implementation.  If we permit ourselves to use unsafe in ordinary packages like encoding/base64 then we are using a trick.  We are telling other people "Go is a safe and fast language" while we use techniques that we don't expect other people to use.  That's not a good way to approach language and library design.  (We have admittedly given ourselves exceptions like strings.Builder, but we try hard to keep those to only the most important parts of the library.)

By all means use your faster version of encoding/base64 yourself, and make it available to other people.  I just don't think that the standard library is the place for code like this.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>