Update patch set 1
Patch Set 1:

Hey Nick, thanks for working on this.

I'd prefer to keep only one copy of TestPackagesFor so that behavior is consistent in 'go vet', 'go test', and 'go list -test'. It's not clear to me the behavior in load.TestPackagesFor is correct, so maybe we can just fix that one by leaving the position on the error.

There is some ambiguity here: for an import error, sometimes we should show the chain of imports, and sometimes we should show the file name and line number. When a package can't be imported (because it's internal, main, has a non-canonical import path, doesn't exist, ...), we currently report the error on the imported package rather than the importer. Perhaps this can be addressed in another issue though.

cc Bryan, Michael because we've been discussing this lately in the context of golang.org/issue/36188.

Patch-set: 1
CC: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 10033 <10033@62eb7196-b449-3ce5-99f1-c037f21e1705>
CC: Gerrit User 17092 <17092@62eb7196-b449-3ce5-99f1-c037f21e1705>