Update patch set 7
Patch Set 7:

(3 comments)

PTAL

I don't feel comfortable not putting the new flag and call to the SessionResetter under the mutex.

That being said, I agree that I can't conceive of any steps that would produce a situation where this would cause a race. Both the "need reset" flag and the calls to SessionResetter and ConnectionValidator are both only called in the "conn" and "putConn". But that would break the model where the driverConn guards all calls to the driver's interface.

Patch-set: 7
Reviewer: Gerrit User 5080 <5080@62eb7196-b449-3ce5-99f1-c037f21e1705>