Update patch set 2
Patch Set 2:

Thanks, I tried this out and thought about it a bit more.

I think the error messages for imports should be the same (or nearly so) for 'go vet' and 'go test'. If a test directly imports a package that can't be imported for some reason, both commands should show a file and line number. If a test transitively imports a package with an error, both commands should show a stack. So it should be possible to fix this in load.TestPackagesFor without any special case in internal/test.

Perhaps the important change is this: in load.TestPackagesFor, if the error is on a package directly imported by the test package, we should show a position. Otherwise (if the error is in a transitively imported package), we should show a stack.

    if len(p1.DepsErrors) > 0 {
      perr := p1.DepsErrors[0]
      direct := len(perr.ImportStack) >= 2 && perr.ImportStack[len(perr.ImportStack)-2] == p1.ImportPath
      if !direct {
        perr.Pos = "" // show full import stack
      }
      err = perr
      break
    }

For reference, the stack printing behavior was introduced in CL 12176 as a fix for golang.org/issue/9558.

Also, I've added a comment to golang.org/issues/26909 about the ambiguity.

Added matloob to reviewers as well.

Patch-set: 2
Reviewer: Gerrit User 10033 <10033@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 17092 <17092@62eb7196-b449-3ce5-99f1-c037f21e1705>