Update patch set 5
Patch Set 5:

(1 comment)

Thanks everyone! I pushed up a new patch set with a buttload of new tests that poke at several different combinations of this problem.

The new behavior isn't perfect, but is at least a lot better. In particular, the code in load.PackagesForBuild is short-circuiting a lot of the smart-stack-trace-printing logic that this code is trying to do.

Patch-set: 5
Reviewer: Gerrit User 34965 <34965@62eb7196-b449-3ce5-99f1-c037f21e1705>