cmd/compile: add more arm64 bitfield simplifications
UBFX and SBFX already zero/sign extend the result.
Further zero/sign extensions are thus unnecessary as long as
they will leave the top bits unaltered.
For unsigned extension, that means that no top bits are populated
by the UBFX.
For signed extension, that means that the top bits will match the
penultimate bit, which is guaranteed as long as that bit is
part of the SBFX sign extension.

These rules collectively trigger 78 times compiling std+cmd.

Change-Id: Ib34d825ff5f42885cc876cafb1a1952b0f88921b