Update patch set 1
Patch Set 1:

(1 comment)

I'm not seeing a race here.  It's OK that there is no lock on the test of .go12 in the isGo12 method.  The isGo12 method calls go12Init.  go12Init ensures that .go12 is initialized.  After go12Init has been called, we know that .go12 will never change.  So after it has been called it's OK to access the .go12 field without holding the lock.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>