Update patch set 2
Patch Set 2:

> Since the test is just running a program that calls into go/build directly, it makes more sense for it to be a unit test in go/build.

Agreed, that is the conclusion I've come to as well after trying to do this via the testscript framework.

> The testscript framework doesn't seem add value here.

The testscript framework makes it easy to create a package outside of standard library with the desired properties (a "// +build ignore" constraint) that build.Import can target without needing to use the internet. I expect this is a bit harder to accomplish manually.

Patch-set: 2
Reviewer: Gerrit User 6005 <6005@62eb7196-b449-3ce5-99f1-c037f21e1705>