cmd/compile: add rewrite to avoid MOV{L,Q}{i2f,f2i}
Possibly for a 1.14 dot release; it's not large, it mitigates
a Darwin/MacOS slowdown from preemptive scheduling (and also
from profiling).

A Darwin bug leaves the upper halves of YMM registers not-clean
after return from a signal handler and this consequently slows
down floating point operations, but especially those involving i2f
and f2i. Adding VZEROUPPER in the injected-call code also helps
with this, but some loops are still not preemptible and hence
run slowly because VZEROUPPER does not happen if the loop cannot
be preempted.

This CL mitigates that problem in a way that is not harmful
on other platforms (it is probably even slightly desirable; a
load and convert are replace with a single load).

This CL includes a commented-out failed attempt at improving
Dasum performance; it is very plausible that this should have
worked, but it did not, even when the compiler was modified
to also hoist the floating point constant load out of the loop.

Updates #37174.

Change-Id: I2f520e01a9df6e0bc0303d4c079b072162d8a48d