cmd/compile: add rewrite to avoid MOV{L,Q}{i2f,f2i}
Possibly for 1.14 or a dot release; it's not large, it mitigates
a Darwin/MacOS slowdown from preemptive scheduling (and also
from profiling).

A Darwin bug leaves the upper halves of YMM registers not-clean
after return from a signal handler (it needs a VZEROUPPER before
it restores the non-zero ones) and this consequently slows down
floating point operations, but especially those involving i2f
and f2i.

This CL mitigates that problem in a way that is not harmful
on other platforms (it is probably even slightly desirable).

This is a mitigation, not a fix, there are known cases
that it does not address (e.g., Dasum in gonum) as well as
floating point code in assembly language.

This CL includes the commented-out fix for Dasum; it is not
a fix until the constant load is properly hoisted out of a loop.

Updates #37174.

Change-Id: I2f520e01a9df6e0bc0303d4c079b072162d8a48d