cmd/compile: add rewrite to avoid MOVLi2f
On Darwin, signals delivery near/at a MOVLi2f instruction
executes astonishingly slowly.  To avoid this, try to elide
the register-register move into a source load instruction, if
that is possible.

I don't think this is a complete fix; there are probably instances
that matter that are not addressed by this simple rewrite.

Updates #37174.

Change-Id: I2f520e01a9df6e0bc0303d4c079b072162d8a48d