obj/x86: nop-only, align functions and jumps for better performance
This is a simplified version of
https://go-review.googlesource.com/c/go/+/206837
that only inserts nops.

This appears to produce slightly slower (+0.2% time) binaries (hard to
tell benchmarking on a laptop) but doesn't cost as much compile time
(again, hard to tell, but larger differences than 0.2%).

name \ build-user-time/op  Go-1.13  Go-1.14-intel  Go-1.14-nop  Go-1.14
[Geo mean]                 4.66s    4.93s          4.75s        4.67s

Resulting binaries usually have the same size; those that did not,
I traced the problem to changes in pclntab (induced by different
distributions of nops or pads).

DO NOT SUBMIT.  I still need to review this code, but this may be
useful for diagnosing performance bugs.

Updates #37121.

Change-Id: Ib9574fcf363d9253a920775968b10b152fcc6a48